fix: adding 'noopener' property when opening a link in a new page
It was possible to verify that, in the chats of attendants and consumers, the button that opens the
sent file does not use the 'noopener' property. The absence of this property allows, if the uploaded
file is compromised, the attacker executes a command that changes the window.opener.location of the
Blip-desk application, thus being able to redirect the attendant to a malicious website.